requirements.txt
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
tenacity>=8.2.2
colorama>=0.4.6
jsonschema>=4.17.3
psutil>=5.9.0
distro>=1.8.0 